+++ Arena/ArenaOnlineGameModes/TeamBattle/TeamBattle.cs
@@ -633,16 +633,12 @@ namespace RainMeadow.Arena.ArenaOnlineGameModes.TeamBattle
             OnlinePlayer player
         )
         {
-
-            if (base.AddIcon(arena, display, owner, customization, player) != "")
-            {
-                return base.AddIcon(arena, display, owner, customization, player);
-            }
+            string arenaIcon = base.AddIcon(arena, display, owner, customization, player);
+            if (arenaIcon != "")
+                return arenaIcon;
 
             if (OnlineManager.lobby.clientSettings.TryGetValue(key: player, out _) == false)
-            {
                 return "";
-            }
 
             if (
                 OnlineManager